% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_mfa_clusterwise.R
\name{block_update_cluster}
\alias{block_update_cluster}
\title{Update a single block's loadings using gradient descent or Adam}
\usage{
block_update_cluster(
  s,
  V_list,
  Sadj,
  row_indices_list,
  Vbig_begin_iter,
  grad_fun_list,
  lambda,
  nsteps_inner,
  optimizer,
  learning_rate,
  global_step,
  m_list = NULL,
  V2_list = NULL,
  beta1,
  beta2,
  adam_epsilon,
  tol_inner,
  verbose,
  iter,
  ncomp_block
)
}
\description{
Update a single block's loadings using gradient descent or Adam
}
\keyword{internal}
