% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_mfa_clusterwise.R
\name{cluster_graph}
\alias{cluster_graph}
\title{Build a cluster graph from spatial coordinates}
\usage{
cluster_graph(coords_list, adjacency_opts = list())
}
\arguments{
\item{coords_list}{List of coordinate matrices (each k_s x 3)}

\item{adjacency_opts}{Options for adjacency construction}
}
\value{
Sparse adjacency matrix
}
\description{
Build a cluster graph from spatial coordinates
}
\keyword{internal}
