% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muscal-package.R
\docType{package}
\name{muscal-package}
\alias{muscal}
\alias{muscal-package}
\title{muscal: Core Multivariate Statistical Analyses and Utilities}
\description{
Provides core 'French school' multivariate analysis methods including Multiple Factor Analysis (MFA), Penalized MFA, STATIS, variants of Correspondence Analysis (BADA, BAMFA), and Generalized Procrustes Analysis. Also includes essential utility functions for data simulation, testing, and processing potentially used by related statistical packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bbuchsbaum.github.io/muscal/}
  \item \url{https://github.com/bbuchsbaum/muscal}
  \item Report bugs at \url{https://github.com/bbuchsbaum/muscal/issues}
}

}
\author{
\strong{Maintainer}: Bradley Buchsbaum \email{bradley.buchsbaum@gmail.com}

}
\keyword{internal}
