% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_mfa.R
\name{penalized_mfa_engine}
\alias{penalized_mfa_engine}
\title{Core engine for penalized MFA optimization}
\usage{
penalized_mfa_engine(
  Xp,
  ncomp,
  lambda,
  penalty_method,
  max_iter,
  nsteps_inner,
  learning_rate,
  optimizer,
  beta1,
  beta2,
  adam_epsilon,
  tol_obj,
  tol_inner,
  verbose,
  compute_consensus
)
}
\description{
Core engine for penalized MFA optimization
}
\keyword{internal}
