% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mfa.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Plot Block Coverage Map}
\usage{
plot_coverage(x, ...)
}
\arguments{
\item{x}{A fitted `linked_mfa` object.}

\item{...}{Additional arguments.}
}
\value{
A ggplot object or base R plot.
}
\description{
Heatmap showing which observations are present in which blocks.
}
