% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mfa.R
\name{plot_variance}
\alias{plot_variance}
\alias{plot_variance.covstatis}
\alias{plot_variance.bada}
\alias{plot_variance.penalized_mfa}
\title{Plot Variance Explained (Scree Plot)}
\usage{
plot_variance(x, type = c("bar", "line"), ...)

plot_variance.covstatis(x, type = c("bar", "line"), ...)

plot_variance.bada(x, type = c("bar", "line"), ...)

plot_variance.penalized_mfa(x, type = c("bar", "line"), ...)
}
\arguments{
\item{x}{A fitted `mfa` or `linked_mfa` object.}

\item{type}{`"bar"` for bar chart, `"line"` for scree plot with cumulative.}

\item{...}{Additional arguments.}
}
\value{
A ggplot object or base R plot.
}
\description{
Shows the proportion of variance captured by each component.
}
