% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_mfa.R
\name{prepare_block_preprocessors}
\alias{prepare_block_preprocessors}
\title{Prepare preprocessors and data for Penalized MFA}
\usage{
prepare_block_preprocessors(
  data_list,
  preproc_arg,
  check_consistent_ncol = TRUE
)
}
\arguments{
\item{check_consistent_ncol}{whether to check for consistent number of columns post-preprocessing.}

\item{data}{list of data blocks}

\item{preproc}{a preprocessor object or list of them}
}
\value{
a list containing the preprocessed data `Xp`, the fitted preprocessors `proclist`,
        and the post-preprocessing dimension `p_post`.
}
\description{
Prepare preprocessors and data for Penalized MFA
}
\keyword{internal}
