% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_mfa.R
\name{print.penalized_mfa}
\alias{print.penalized_mfa}
\title{Print Method for Penalized MFA Objects}
\usage{
\method{print}{penalized_mfa}(x, ...)
}
\arguments{
\item{x}{A `penalized_mfa` object}

\item{...}{Additional parameters (unused)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for Penalized MFA Objects
}
