% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covstatis.R
\name{roi_map}
\alias{roi_map}
\title{Extract ROI map for a compromise dimension}
\usage{
roi_map(x, d)
}
\arguments{
\item{x}{A fitted `covstatis` object.}

\item{d}{The dimension to extract.}
}
\value{
A numeric vector representing the ROI map for the specified dimension.
}
\description{
Extract ROI map for a compromise dimension
}
