% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_mfa_clusterwise.R
\name{spatial_constraints}
\alias{spatial_constraints}
\title{Simple spatial constraints function (creates k-NN adjacency matrix)}
\usage{
spatial_constraints(coords_list, k_nn = 6, nblocks = NULL, ...)
}
\arguments{
\item{coords_list}{List of coordinate matrices (each k_s x 3)}

\item{k_nn}{Number of nearest neighbors (default 6)}

\item{nblocks}{Number of blocks (ignored, for compatibility)}
}
\value{
Sparse adjacency matrix
}
\description{
Simple spatial constraints function (creates k-NN adjacency matrix)
}
\keyword{internal}
