% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_mfa_clusterwise.R
\name{validate_pmfa_args}
\alias{validate_pmfa_args}
\title{Validate arguments for penalized MFA functions}
\usage{
validate_pmfa_args(
  data_list,
  coords_list = NULL,
  ncomp,
  lambda,
  max_iter,
  nsteps_inner,
  learning_rate,
  optimizer,
  beta1,
  beta2,
  adam_epsilon,
  tol_obj,
  tol_inner,
  verbose,
  memory_budget_mb
)
}
\description{
Validate arguments for penalized MFA functions
}
\keyword{internal}
