% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covstatis.R
\name{plot_subjects}
\alias{plot_subjects}
\title{Plot subject clouds in RV and ROI spaces}
\usage{
plot_subjects(x, dim = c(1, 2))
}
\arguments{
\item{x}{A fitted `covstatis` object.}

\item{dim}{A numeric vector of length 2 specifying the dimensions to plot.}
}
\description{
Plot subject clouds in RV and ROI spaces
}
